refactor: The default value of cssEllipsis becomes false, unit test add (#1039)
