+++ components/Typography/interface.tsx
@@ -144,9 +144,10 @@ export type TypographyTextProps = CommonProps;
  */
 export type EllipsisConfig = {
   /**
-   * @zh 自动溢出省略（只支持字符串），在简单的单行省略情况下，会默认使用 css 处理省略，避免复杂计算。
+   * @zh 自动溢出省略（只支持字符串），在大量使用情况下建议开启提高性能。
    * @en Automatic overflow omission (only strings are supported). In the case of simple single-line, css will be used by default to handle ellipsis to avoid complicated calculations.
-   * @defaultValue true
+   * @defaultValue false
+   * @version `2.36.0` 将默认值改为 `false` 并支持多行CSS省略。
    */
   cssEllipsis?: boolean;
   /**
